[RF] Move BufferManager into RooBatchCompute
This is only the first step. The next step is to move it into the
dynamically loaded part of RooBatchCompute, and have different CPU and
GPU implementations.
