+++ roofit/roofitcuda/CMakeLists.txt
@@ -14,7 +14,7 @@ ROOT_LINKER_LIBRARY(RooFitCuda ${cuda_sources} src/CudaHelpers.cxx TYPE SHARED)
 set_target_properties(RooFitCuda PROPERTIES CUDA_RUNTIME_LIBRARY Shared)
 
 if(cuda)
-  target_compile_definitions(RooFitCuda PRIVATE ROOFIT_CUDA)
+    target_compile_definitions(RooFitCuda PUBLIC ROOFIT_CUDA)
 endif()
 
 ROOT_INSTALL_HEADERS()
